requirements.txt
streamlit
python-dotenv
langchain-community
ollama